+++ src/server/devices/baseDeviceInterfaces/iEnergyManager.ts
@@ -4,6 +4,10 @@ import { iBaseDevice } from './iBaseDevice';
 export interface iEnergyManager extends iBaseDevice {
   excessEnergy: number;
 
+  readonly injectingWattage: number;
+  readonly drawingWattage: number;
+  readonly selfConsumingWattage: number;
+
   addExcessConsumer(device: iExcessEnergyConsumer): void;
 
   recalculatePowerSharing(): void;